refactor: reorganize audit system and general fixes
Main changes:

- Audit refactoring: moved from logger_config.py to utils/audit.py
  - New audit() function returns filtered list and product count
  - Output in JSON format (audit_products.json) in logs/files/
  - Simplified code in terracast.py

- Bug fixes:
  - product_processor.py: fixed processed_files merge (append vs overwrite)
  - product_processor.py: fixed CONFIG['output_vis'] usage for vis_dir
  - html_update.py: added num_files > 0 check to prevent IndexError

- Legacy code cleanup:
  - Removed _legacy/config.py (~4500 unused lines)
  - Removed logs/log.py (unused)
  - Removed AUDIT level code from logger_config.py

- Updated dependencies (requirements.txt):
  - Removed: watchdog
  - Added: pyspectral, pyhdf

- Minor fixes in processing scripts

+++ scripts/g1X_false_color_fdk.py
@@ -225,11 +225,12 @@ data_ch13_original = data_ch13
 IRmin = 89.62
 IRmax = 341.27
 # Anything that is below the min or greater than max, keep min and max
+data_ch13 = data_ch13.astype(int)
 data_ch13[data_ch13 > IRmax] = IRmax
 # Convert to 0 - 255
 data_ch13 = ((data_ch13 - IRmin) / (IRmax - IRmin)) * 255
 # Convert to int
-data_ch13 = data_ch13.astype(int)
+
 
 #print("Reading the False Color Look Up Table...")
